requirements.txt
fastapi>=0.109.0,<1
uvicorn[standard]>=0.27.0,<1
sqlalchemy>=2.0.25,<3
psycopg2-binary>=2.9.9,<3
passlib[bcrypt]>=1.7.4,<2
bcrypt<4.1
python-dotenv>=1.0.0,<2
PyJWT>=2.8.0,<3
pydantic-settings>=2.1.0,<3
